fix warmup sql
