+++ src/analytics/private/constant.ts
@@ -30,16 +30,6 @@ export enum WorkflowStatus {
 
 export const DYNAMODB_TABLE_INDEX_NAME = 'status_timestamp_index';
 
-export const REDSHIFT_EVENT_TABLE_NAME = 'event';
-export const REDSHIFT_EVENT_PARAMETER_TABLE_NAME = 'event_parameter';
-export const REDSHIFT_USER_TABLE_NAME = 'user';
-export const REDSHIFT_ITEM_TABLE_NAME = 'item';
-
-export const REDSHIFT_TABLE_NAMES = [
-  REDSHIFT_EVENT_TABLE_NAME, REDSHIFT_EVENT_PARAMETER_TABLE_NAME,
-  REDSHIFT_ITEM_TABLE_NAME, REDSHIFT_USER_TABLE_NAME,
-];
-
 export const REDSHIFT_DUPLICATE_DATE_INTERVAL = 3; // Days
 
 export const SP_SCAN_METADATA = 'scan_metadata';
